==> 1.1 Lematizacion términos y definiciones.ipynb <==
import pandas as pd

file_path = 'terminos_definiciones.xlsx'  # Reemplaza con la ruta a tu archivo
df = pd.read_excel(file_path)
para_lematizar = df.columns[0]

import pandas as pd

# Ruta al archivo de Excel
archivo_excel = 'terminos_definiciones.xlsx'

# Leer la hoja llamada 'Hoja1', usando la primera fila como cabecera, leyendo solo columnas A a C, saltando la primera fila y leyendo 10 filas.
df = pd.read_excel(archivo_excel, sheet_name='Hoja1', header=0, usecols="A:B", skiprows=None, nrows=None)

# Mostrar las primeras filas del DataFrame
print(df.head())


terminos = set()
for termino in df['Término']:
    terminos.add(termino)

import requests
def lematiza(fila):
    cadena_entrada = fila["Definiciones"]
    archivo = open("archivo.txt","w",encoding="utf-8")
    archivo.write(cadena_entrada)
    archivo.close()
    files = {'file': open("archivo.txt", 'rb')}
    #Parámetros
    params = {'outf': 'tagged', 'format': 'json'}
    #Enviar petición
    url = "http://www.corpus.unam.mx/servicio-freeling/analyze.php"
    r = requests.post(url, files=files, params=params)
    #Convertir de formato json
    obj = r.json()
    #Ejemplo, obtener todos los lemas
    #print(obj)
    cadena = ""
    for sentence in obj:
      for word in sentence:
        cadena = cadena + " "+ word["lemma"]
    return pd.Series([cadena])

df[['Def_lema']] = df.apply(lematiza, axis=1)

df.to_excel('termino_definiciones_lematizados.xlsx', index=False)

import re
def limpiar_espacios_multipalabras(cadena):
    # Eliminar espacios después de paréntesis de apertura y antes de paréntesis de cierre
    cadena = re.sub(r'\(\s+', '(', cadena)  # Quita espacios después de '('
    cadena = re.sub(r'\s+\)', ')', cadena)  # Quita espacios antes de ')'
    # Eliminar espacios al inicio y al final de la cadena
    cadena = cadena.strip()
    cadena = cadena.replace(" ", "_")
    return cadena

import os

terminos = {}
for archivo in os.listdir("terminos"):
    original = open("terminos/"+archivo)
    cadena_original = original.readline()
    traducido = open("terminos_lematizados/"+archivo)
    cadena_traducida = traducido.readline()
    print(cadena_original+":"+limpiar_espacios_multipalabras(cadena_traducida))
    terminos[cadena_original] = limpiar_espacios_multipalabras(cadena_traducida)

def lema_termino(fila):
    term = fila["Término"]
    nuevo_term = terminos[term]
    return pd.Series([nuevo_term])


df[['term_lema']] = df.apply(lema_termino, axis=1)

df.to_excel('termino_definiciones_lematizados.xlsx', index=False)

import nltk
import re
from nltk.tokenize import word_tokenize
from nltk.corpus import stopwords
nltk.download('punkt')
nltk.download('stopwords')
vacias = stopwords.words('spanish')nlp = spacy.load('es_core_news_sm')

import spacy
nlp = spacy.load('es_core_news_sm')

def remueve_stop(fila):
    contenido = fila["Def_lema"]
    contenido = contenido.strip()
    patron = r'[^\w\s]'
    contenido = re.sub(patron, '',contenido)
    doc = nlp(contenido)
    palabras_sin_stopwords = [token.text for token in doc if not token.is_stop]
    sinstop = ' '.join(palabras_sin_stopwords)
    token = word_tokenize(sinstop)
    for palabra in vacias:
        token = [s for s in token if s != palabra]
    token = [s for s in token if s != " "]
    cadena = " ".join(token)
    return pd.Series([cadena])

df[['def_lema_limpia']] = df.apply(remueve_stop, axis=1)

df.to_excel('termino_definiciones_lematizados.xlsx', index=False)
