==> 2. Lematización corpus.ipynb <==
import re
def limpiar_espacios(cadena):
    # Eliminar espacios después de paréntesis de apertura y antes de paréntesis de cierre
    cadena = re.sub(r'\(\s+', '(', cadena)  # Quita espacios después de '('
    cadena = re.sub(r'\s+\)', ')', cadena)  # Quita espacios antes de ')'
    # Eliminar espacios al inicio y al final de la cadena
    cadena = cadena.strip()
    return cadena

import re

def limpiar_texto(nombre_archivo):
    # Leer el contenido del archivo
    with open(nombre_archivo, 'r', encoding='utf-8') as archivo:
        contenido = archivo.read()
    # Expresión regular para encontrar secuencias de escape
    escape_sequence_pattern = r'\\.'

    # Usar re.sub para quitar las secuencias de escape
    contenido = re.sub(escape_sequence_pattern, ' ', contenido)

    # Eliminar signos de puntuación
    contenido = re.sub(r'[^\w\s]', '', contenido)

    # Eliminar secuencias de escape \r\n
    #contenido = contenido.replace('\n', '')
    #contenido = contenido.replace('\t', '')
    #contenido = contenido.replace('\r', '')

    nombre_archivo_limpio = nombre_archivo
    nombre_archivo_limpio = nombre_archivo_limpio.replace('corpus/cientifico/', '')
    with open("corpus/cientifico_limpio/"+nombre_archivo_limpio, 'w', encoding='utf-8') as archivo_limpio:
        archivo_limpio.write(contenido)

    print(f"Texto limpio guardado en: {nombre_archivo_limpio}")


for archivo in os.listdir("corpus/cientifico/"):
    limpiar_texto("corpus/cientifico/"+archivo)

import os
import requests

for archivo in os.listdir("corpus/cientifico_limpio/"):
    files = {'file': open("corpus/cientifico_limpio/"+archivo, 'rb')}
    #Parámetros
    params = {'outf': 'tagged', 'format': 'json'}
    #Enviar petición
    url = "http://www.corpus.unam.mx/servicio-freeling/analyze.php"
    r = requests.post(url, files=files, params=params)
    #Convertir de formato json
    obj = r.json()
    #Ejemplo, obtener todos los lemas
    #print(obj)
    cadena = ""
    for sentence in obj:
      for word in sentence:
        cadena = cadena + " "+ word["lemma"]
    cadena = cadena.lower()
    archivo_salida = open("corpus/cientifico_lematizado/"+archivo,"w",encoding="utf-8")
    archivo_salida.write(cadena)
    archivo_salida.close()

import re

def limpiar_texto(nombre_archivo):
    # Leer el contenido del archivo
    with open(nombre_archivo, 'r', encoding='utf-8') as archivo:
        contenido = archivo.read()
    # Expresión regular para encontrar secuencias de escape
    escape_sequence_pattern = r'\\.'

    # Usar re.sub para quitar las secuencias de escape
    contenido = re.sub(escape_sequence_pattern, ' ', contenido)

    # Eliminar signos de puntuación
    contenido = re.sub(r'[^\w\s]', '', contenido)

    # Eliminar secuencias de escape \r\n
    #contenido = contenido.replace('\n', '')
    #contenido = contenido.replace('\t', '')
    #contenido = contenido.replace('\r', '')

    # Escribir el texto limpio en un nuevo archivo
    nombre_archivo_limpio = nombre_archivo
    nombre_archivo_limpio = nombre_archivo_limpio.replace('corpus/informativo/', '')
    with open("corpus/informativo_limpio/"+nombre_archivo_limpio, 'w', encoding='utf-8') as archivo_limpio:
        archivo_limpio.write(contenido)

    print(f"Texto limpio guardado en: {nombre_archivo_limpio}")

for archivo in os.listdir("corpus/informativo/"):
    limpiar_texto("corpus/informativo/"+archivo)

import os
import requests

for archivo in os.listdir("corpus/informativo_limpio/"):
    files = {'file': open("corpus/informativo_limpio/"+archivo, 'rb')}
    #Parámetros
    params = {'outf': 'tagged', 'format': 'json'}
    #Enviar petición
    url = "http://www.corpus.unam.mx/servicio-freeling/analyze.php"
    r = requests.post(url, files=files, params=params)
    #Convertir de formato json
    obj = r.json()
    #Ejemplo, obtener todos los lemas
    #print(obj)
    cadena = ""
    for sentence in obj:
      for word in sentence:
        cadena = cadena + " "+ word["lemma"]
    cadena = cadena.lower()
    archivo_salida = open("corpus/informativo_lematizado/"+archivo,"w",encoding="utf-8")
    archivo_salida.write(cadena)
    archivo_salida.close()

import re

def limpiar_texto(nombre_archivo):
    # Leer el contenido del archivo
    with open(nombre_archivo, 'r', encoding='utf-8') as archivo:
        contenido = archivo.read()
    # Expresión regular para encontrar secuencias de escape
    escape_sequence_pattern = r'\\.'

    # Usar re.sub para quitar las secuencias de escape
    contenido = re.sub(escape_sequence_pattern, ' ', contenido)

    # Eliminar signos de puntuación
    contenido = re.sub(r'[^\w\s]', '', contenido)

    # Eliminar secuencias de escape \r\n
    #contenido = contenido.replace('\n', '')
    #contenido = contenido.replace('\t', '')
    #contenido = contenido.replace('\r', '')

    # Escribir el texto limpio en un nuevo archivo
    nombre_archivo_limpio = nombre_archivo
    nombre_archivo_limpio = nombre_archivo_limpio.replace('corpus/wikipedia/', '')
    with open("corpus/wikipedia_limpio/"+nombre_archivo_limpio, 'w', encoding='utf-8') as archivo_limpio:
        archivo_limpio.write(contenido)

    print(f"Texto limpio guardado en: {nombre_archivo_limpio}")

for archivo in os.listdir("corpus/wikipedia/"):
    limpiar_texto("corpus/wikipedia/"+archivo)

import os
import requests

for archivo in os.listdir("corpus/wikipedia_limpio/"):
    files = {'file': open("corpus/wikipedia_limpio/"+archivo, 'rb')}
    #Parámetros
    params = {'outf': 'tagged', 'format': 'json'}
    #Enviar petición
    url = "http://www.corpus.unam.mx/servicio-freeling/analyze.php"
    r = requests.post(url, files=files, params=params)
    #Convertir de formato json
    obj = r.json()
    #Ejemplo, obtener todos los lemas
    #print(obj)
    cadena = ""
    for sentence in obj:
      for word in sentence:
        cadena = cadena + " "+ word["lemma"]
    cadena = cadena.lower()
    archivo_salida = open("corpus/wikipedia_lematizado/"+archivo,"w",encoding="utf-8")
    archivo_salida.write(cadena)
    archivo_salida.close()
