==> 4. Grafo coocurrencias.ipynb <==
!pip install text2graphapi

archivo = open('csmx.txt',encoding="utf-8")
contenido = archivo.read()

import nltk
import re
from nltk.tokenize import word_tokenize
from nltk.corpus import stopwords
nltk.download('punkt')
nltk.download('stopwords')
vacias = stopwords.words('spanish')

import spacy

def dividir_cadena(cadena):
    # Calcular la longitud de cada parte
    longitud_total = len(cadena)
    longitud_parte = longitud_total // 3
    
    # Calcular los índices de corte
    indice1 = longitud_parte
    indice2 = 2 * longitud_parte
    
    # Dividir la cadena usando slicing
    parte1 = cadena[:indice1]
    parte2 = cadena[indice1:indice2]
    parte3 = cadena[indice2:]
    
    return parte1, parte2, parte3

nlp = spacy.load('es_core_news_sm') 
contenido = contenido.strip()
patron = r'[^\w\s]'
parte1, parte2, parte3 = dividir_cadena(contenido)
partes = []
partes.append(parte1)
partes.append(parte2)
partes.append(parte3)
todo_sinstop = ""
nlp.max_length = len(contenido)
for i in range(0,3):
    contenido = re.sub(patron, '', partes[i])
    doc = nlp(contenido)
    palabras_sin_stopwords = [token.text for token in doc if not token.is_stop]
    parte_sinstop = ' '.join(palabras_sin_stopwords)
    todo_sinstop += parte_sinstop

token = word_tokenize(todo_sinstop)
for palabra in vacias:
    token = [s for s in token if s != palabra]
token = [s for s in token if s != " "]
cadena = " ".join(token)

archivo_salida = open("csmx_lematizado_sin_stop.txt","w",encoding="utf-8")
archivo_salida.write(cadena)
archivo_salida.close()

input_text_docs = [
	{"id": 1, "doc": cadena}
]

from text2graphapi.src.Cooccurrence import Cooccurrence

to_cooccurrence = Cooccurrence(
                graph_type = 'Graph',
                apply_prep = False,
                parallel_exec = True,
                window_size = 10,
                language = 'sp',
                output_format = 'networkx')

output_text_graphs = to_cooccurrence.transform(input_text_docs)
print(output_text_graphs)

import networkx as nx
grafo_csmx = output_text_graphs[0]['graph']
print(grafo_csmx)
print(list(grafo_csmx.edges)[None:1000:None])

nx.write_gexf(grafo_csmx,"grafo_csmx_ventana10.gexf", encoding='utf-8')
