==> 5. Normas.ipynb <==
import networkx as nx
grafo_csmx = nx.read_gexf("grafo_csmx_ventana10.gexf")
peso=nx.get_edge_attributes(grafo_csmx,'freq')


import pandas as pd

import os
import re
def limpiar_espacios(cadena):
    # Eliminar espacios después de paréntesis de apertura y antes de paréntesis de cierre
    cadena = re.sub(r'\(\s+', '(', cadena)  # Quita espacios después de '('
    cadena = re.sub(r'\s+\)', ')', cadena)  # Quita espacios antes de ')'
    # Eliminar espacios al inicio y al final de la cadena
    cadena = cadena.strip()
    return cadena

terminos = {}
for archivo in os.listdir("terminos"):
    original = open("terminos/"+archivo)
    cadena_original = original.readline()
    traducido = open("terminos_lematizados/"+archivo)
    cadena_traducida = traducido.readline()
    print(cadena_original+":"+limpiar_espacios(cadena_traducida))
    terminos[cadena_original] = limpiar_espacios(cadena_traducida)

terminos_lematizados = []
for llave in terminos.keys():
    terminos_lematizados.append(terminos[llave])

def unir_multipalabras(lista_palabras):    
    return [palabra.replace(" ", "_") for palabra in lista_palabras]


terminos_lematizados_unidos = unir_multipalabras(terminos_lematizados)

terminos_lematizados_unidos

directorio = "normas/"
peso=nx.get_edge_attributes(grafo_csmx,'freq')
arreglo_tuplas = []
# Crear un DataFrame vacío con las columnas deseadas
df = pd.DataFrame(columns=['Respuesta', 'Frecuencia', 'Asociación'])
for term in terminos_lematizados_unidos:
    if term in grafo_csmx.nodes():
        suma = 0
        for vecino in grafo_csmx.neighbors(term):
            if vecino != term:
                try:
                    peso_local = int(peso[(vecino,term)])
                    suma += peso_local
                except:
                    peso_local = int(peso[(term,vecino)])
                    suma += peso_local
        comprobacion = 0
        arreglo_tuplas = []
        for vecino in grafo_csmx.neighbors(term):
            if vecino != term:
                try:
                    peso_local = int(peso[(vecino,term)])                
                except:
                    peso_local = int(peso[(term,vecino)])                
                tupla = (vecino,peso_local,float(peso_local/suma))
                comprobacion += float(peso_local/suma)
                arreglo_tuplas.append(tupla)
        print("Término ",term)   
        arreglo_tuplas = sorted(arreglo_tuplas, key=lambda x: x[1],reverse=True)
        #print(arreglo_tuplas)
        #print("El total de la asociación es", comprobacion)        
        #input()
        df = pd.DataFrame(arreglo_tuplas, columns=['Respuesta', 'Frecuencia', 'Asociación'])
        df.to_csv(directorio+term+'.csv',index=False)
