==> 7. Diccionario-solo-terminos-normas.ipynb <==
import pandas as pd
import os
import networkx as nx

def remover_nodos_vacios(grafo):
    vacias = open("removed_words.txt",encoding="utf-8")
    palabras = vacias.readlines()
    for palabra in palabras:
        if palabra in grafo.nodes():
            grafo.remove_node(palabra)
    return grafo

grafo_a = nx.Graph()
grafo_f = nx.Graph()
directorio = 'normas/'
for archivo in os.listdir(directorio):
    datos = pd.read_csv(directorio+archivo)
    estimulo = archivo[:archivo.find(".csv")]
    for respuesta, frecuencia,asociacion in zip(datos["Respuesta"],datos["Frecuencia"],datos["Asociación"]):
        grafo_f.add_edge(estimulo,respuesta,weight=float(1/int(frecuencia)))
        grafo_a.add_edge(estimulo,respuesta,weight=float(1-int(asociacion)))
print("Removiendo nodos...")
remover_nodos_vacios(grafo_f)
remover_nodos_vacios(grafo_a)
print(grafo_a)
print(grafo_f)

data = pd.read_excel("termino_definiciones_lematizados.xlsx")
corpus_evaluacion = dict()
for termino,definicion in zip(data["term_lema"],data["def_lema_limpia"]):
    if termino in corpus_evaluacion.keys():
        definicion = str(corpus_evaluacion[termino]) + "#" + str(definicion) #las definiciones con el mismo término se agregan a la misma llave pero separadas por #
    corpus_evaluacion[termino] = definicion

def filtro_grados(candidato):
    if grafo_a.degree[candidato] > promedio:
        return False
    else:
        return True

def conceptos(lista,subconjunto):
    datos = []
    if len(lista) <= 38:
        tope = len(lista)
    else:
        tope = 38
    for x in range(0,tope):
        if str(lista[x][0]) not in subconjunto and float(lista[x][1])>0:
            datos.append(str(lista[x][0]))
    return datos

def conceptos_sin_filtrar(lista):
    datos = []
    for x in range(len(lista)):
        datos.append(str(lista[x][0]))
    return datos

def reducir_grafo(subconjunto,grafo_original):
    grafo_reducido = nx.Graph()
    nuevos_nodos = []
    nuevos_nodos.extend(subconjunto)
    for nodo in subconjunto:
      vecinos = grafo_original[nodo]
      for vecino in vecinos:
        nuevos_nodos.append(vecino)
    grafo_reducido = grafo_original.subgraph(nuevos_nodos)
    return grafo_reducido

def precision(k,arreglo,concepto):
    tope = k
    encontrado = 0
    if k>len(arreglo):
        tope = len(arreglo)
    for i in range(0,tope):
        if arreglo[i] == concepto.strip():
            encontrado = 1
    return encontrado

import os.path

def revisa_pesos_normas(palabras_en_grafo,candidatos):
    directorio = 'normas/'
    pesos_frec = []
    pesos_asoc = []
    for candidato in candidatos:
        suma_frec = 0
        suma_asoc = 0
        for palabra in palabras_en_grafo:
            if os.path.isfile(directorio+candidato+".csv") == True:
                datos = pd.read_csv(directorio+candidato+".csv")
                #print(datos["Respuesta"])
                if palabra in datos["Respuesta"].values:
                    row_num = datos[datos['Respuesta'] == palabra].index
                    frec = float(datos.iloc[row_num]["Frecuencia"])
                    asoc = float(datos.iloc[row_num]["Asociación"])
                    suma_frec = suma_frec + frec
                    suma_asoc = suma_asoc + asoc
        pesos_frec.append(suma_frec)
        pesos_asoc.append(suma_asoc)
    return pesos_frec, pesos_asoc

from operator import itemgetter

def diccionario_btc(grafo,definicion,termino):
    texto = str(definicion)
    texto = texto.lower()
    tokens = texto.split(" ")
    subconjunto_lemas = []
    for palabra in tokens:
        if palabra in grafo.nodes() and palabra != '':
            subconjunto_lemas.append(palabra)
    if len(subconjunto_lemas) > 0:
        nuevo_grafo = reducir_grafo(subconjunto_lemas,grafo)
        resultados_asociacion = nx.betweenness_centrality_subset(nuevo_grafo, subconjunto_lemas, subconjunto_lemas, normalized=True, weight='weight')
        encontrados = sorted(resultados_asociacion.items(),key=itemgetter(1),reverse=True)[0:37]
        return conceptos(encontrados,subconjunto_lemas), subconjunto_lemas
    else:
        return [],[]

def rendimiento_mdbl_btc():
    print("hola")
    p_1_a_no = p_1_a = 0
    p_3_a_no = p_3_a = 0
    p_5_a_no = p_5_a = 0
    p_10_a_no = p_10_a = 0
    suma_normal = suma_no = 0
    #total = len(corpus_evaluacion.keys())
    #total = 0
    i=0
    total = 0
    directorio = 'normas/'
    lista_normas = os.listdir(directorio)
    archivos_sin_extension = [os.path.splitext(archivo)[0] for archivo in lista_normas]
    for termino in corpus_evaluacion.keys():
        definiciones = str(corpus_evaluacion[termino]).split("#")
        if termino in archivos_sin_extension:
            for definicion in definiciones:
                candidatos,palabras_en_grafo = diccionario_btc(grafo_a,definicion,termino)
                i = i + 1
                total = total + 1
                print(i)
                peso_tot_frec, peso_tot_asoc = revisa_pesos_normas(palabras_en_grafo,candidatos)
                """print("Término:",termino)
                print("Palabras de la definición:",palabras_en_grafo)
                print("Candidatos:",candidatos)
                print("Pesos frecuencia:",peso_tot_frec)
                print("Pesos asociación:",peso_tot_asoc)"""
                p_1_a += precision(1,candidatos,termino)
                p_3_a += precision(3,candidatos,termino)
                p_5_a += precision(5,candidatos,termino)
                p_10_a += precision(10,candidatos,termino)
                try:
                    pos_normal = candidatos.index(termino) + 1
                except:
                    pos_normal = 0
                nuevo_orden_candidatos = dict(zip(candidatos,peso_tot_asoc))
                encontrados = sorted(nuevo_orden_candidatos.items(),key=itemgetter(1),reverse=True)
                candidatos = conceptos_sin_filtrar(encontrados)
                #print("Candidatos nuevo orden:",candidatos)
                p_1_a_no += precision(1,candidatos,termino)
                p_3_a_no += precision(3,candidatos,termino)
                p_5_a_no += precision(5,candidatos,termino)
                p_10_a_no += precision(10,candidatos,termino)
                try:
                    pos_no = candidatos.index(termino) + 1
                except:
                    pos_no = 0
                if pos_normal != 0:
                    suma_normal = suma_normal + (1/pos_normal)
                if pos_no != 0:
                    suma_no = suma_no + (1/pos_no)
    print("Total de deficiones:",total)
    print("Asociación")
    print("p@1:",p_1_a/total,"p@3",p_3_a/total,"p@5",p_5_a/total,"p@10",p_10_a/total)
    print("MRR:",(1/total)*(suma_normal))
    print("Nuevo orden")
    print("p@1:",p_1_a_no/total,"p@3",p_3_a_no/total,"p@5",p_5_a_no/total,"p@10",p_10_a_no/total)
    print("MRR:",(1/total)*(suma_no))
    return

import warnings
warnings.filterwarnings("ignore")

rendimiento_mdbl_btc()
